Add apiCall to fetch real data
